+++ Renderer/src/main/java/net/buildabrowser/babbrowser/browser/render/content/common/PaddingUtil.java
@@ -15,14 +15,14 @@ public final class PaddingUtil {
     LayoutContext layoutContext, ElementBox childBox, LayoutConstraint referenceConstraint
   ) {
     ActiveStyles styles = childBox.activeStyles();
-    int topPadding = computePadding(styles.getProperty(CSSProperty.PADDING_TOP), layoutContext, childBox, referenceConstraint);
-    int bottomPadding = computePadding(styles.getProperty(CSSProperty.PADDING_BOTTOM), layoutContext, childBox, referenceConstraint);
-    int leftPadding = computePadding(styles.getProperty(CSSProperty.PADDING_LEFT), layoutContext, childBox, referenceConstraint);
-    int rightPadding = computePadding(styles.getProperty(CSSProperty.PADDING_RIGHT), layoutContext, childBox, referenceConstraint);
+    float topPadding = computePadding(styles.getProperty(CSSProperty.PADDING_TOP), layoutContext, childBox, referenceConstraint);
+    float bottomPadding = computePadding(styles.getProperty(CSSProperty.PADDING_BOTTOM), layoutContext, childBox, referenceConstraint);
+    float leftPadding = computePadding(styles.getProperty(CSSProperty.PADDING_LEFT), layoutContext, childBox, referenceConstraint);
+    float rightPadding = computePadding(styles.getProperty(CSSProperty.PADDING_RIGHT), layoutContext, childBox, referenceConstraint);
     childBox.dimensions().setComputedPadding(topPadding, bottomPadding, leftPadding, rightPadding);
   }
 
-  private static int computePadding(
+  private static float computePadding(
     CSSValue property, LayoutContext layoutContext, ElementBox childBox, LayoutConstraint referenceConstraint
   ) {
     LayoutConstraint constraint = SizingUtil.evaluateBaseSize(layoutContext, referenceConstraint, property);
