+++ Renderer/src/main/java/net/buildabrowser/babbrowser/browser/render/content/common/fragment/BoxFragment.java
@@ -11,7 +11,7 @@ public class BoxFragment extends LayoutFragment {
   private LayoutFragment parentFragment;
 
   public BoxFragment(
-    int width, int height,
+    float width, float height,
     ElementBox box, BoxPainter painter
   ) {
     super(width, height);
@@ -32,46 +32,46 @@ public class BoxFragment extends LayoutFragment {
   }
 
   @Override
-  public int contentX() {
-    int[] border = box.dimensions().getComputedBorder();
-    int[] padding = box.dimensions().getComputedPadding();
+  public float contentX() {
+    float[] border = box.dimensions().getComputedBorder();
+    float[] padding = box.dimensions().getComputedPadding();
     return borderX() + border[2] + padding[2];
   }
 
   @Override
-  public int contentY() {
-    int[] border = box.dimensions().getComputedBorder();
-    int[] padding = box.dimensions().getComputedPadding();
+  public float contentY() {
+    float[] border = box.dimensions().getComputedBorder();
+    float[] padding = box.dimensions().getComputedPadding();
     return borderY() + border[0] + padding[0];
   }
 
   @Override
-  public int layerX() {
-    int[] padding = box.dimensions().getComputedPadding();
+  public float layerX() {
+    float[] padding = box.dimensions().getComputedPadding();
     return parentFragment == null ?
       padding[2] :
       parentFragment.layerX() + contentX();
   }
 
   @Override
-  public int layerY() {
-    int[] padding = box.dimensions().getComputedPadding();
+  public float layerY() {
+    float[] padding = box.dimensions().getComputedPadding();
     return parentFragment == null ?
       padding[0] :
       parentFragment.layerY() + contentY();
   }
 
   @Override
-  public int borderWidth() {
-    int[] border = box.dimensions().getComputedBorder();
-    int[] padding = box.dimensions().getComputedPadding();
+  public float borderWidth() {
+    float[] border = box.dimensions().getComputedBorder();
+    float[] padding = box.dimensions().getComputedPadding();
     return contentWidth() + border[2] + border[3] + padding[2] + padding[3];
   }
 
   @Override
-  public int borderHeight() {
-    int[] border = box.dimensions().getComputedBorder();
-    int[] padding = box.dimensions().getComputedPadding();
+  public float borderHeight() {
+    float[] border = box.dimensions().getComputedBorder();
+    float[] padding = box.dimensions().getComputedPadding();
     return contentHeight() + border[0] + border[1] + padding[0] + padding[1];
   }
 
@@ -79,26 +79,26 @@ public class BoxFragment extends LayoutFragment {
   // WARNING: Does not account for margin collapsing
 
   @Override
-  public int marginX() {
-    int[] margin = box.dimensions().getComputedMargin();
+  public float marginX() {
+    float[] margin = box.dimensions().getComputedMargin();
     return borderX() - margin[2];
   }
 
   @Override
-  public int marginY() {
-    int[] margin = box.dimensions().getComputedMargin();
+  public float marginY() {
+    float[] margin = box.dimensions().getComputedMargin();
     return borderY() - margin[0];
   }
 
   @Override
-  public int marginWidth() {
-    int[] margin = box.dimensions().getComputedMargin();
+  public float marginWidth() {
+    float[] margin = box.dimensions().getComputedMargin();
     return borderWidth() + margin[2] + margin[3];
   }
 
   @Override
-  public int marginHeight() {
-    int[] margin = box.dimensions().getComputedMargin();
+  public float marginHeight() {
+    float[] margin = box.dimensions().getComputedMargin();
     return borderHeight() + margin[0] + margin[1];
   }
