+++ Renderer/src/main/java/net/buildabrowser/babbrowser/browser/render/paint/FontMetrics.java
@@ -2,8 +2,8 @@ package net.buildabrowser.babbrowser.browser.render.paint;
 
 public interface FontMetrics {
   
-  int stringWidth(String text);
+  float stringWidth(String text);
 
-  int fontHeight();
+  float fontHeight();
 
 }
